bot/requirements.txt
aiogram>=3.15
aiohttp>=3.10
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
python-dotenv>=1.0
